#Add validate package command#130
refactoring
