+++ clio/Environment/ClioEnvironment.cs
@@ -2,30 +2,12 @@
 using System.IO;
 using System.Linq;
 using System.Reflection;
-using Creatio.Client;
 
 namespace Clio.UserEnvironment
 {
 	internal class CreatioEnvironment : ICreatioEnvironment
 	{
 		private const string PathVariableName = "PATH";
-		public static string AppUrl {
-			get {
-				if (CreatioEnvironment.IsNetCore) {
-					return Url;
-				} else {
-					return Url + @"/0";
-				}
-			}
-		}
-
-		public static string Url => Settings.Uri; // Необходимо получить из конфига
-
-		public static string GetZipPackageUrl => AppUrl + @"/ServiceModel/PackageInstallerService.svc/GetZipPackages";
-
-		public static string GetEntityModelsUrl => AppUrl + @"/rest/CreatioApiGateway/GetEntitySchemaModels/{0}";
-
-		public static string ApiVersionUrl => AppUrl + @"/rest/CreatioApiGateway/GetApiVersion";
 
 		public static bool IsNetCore => Settings.IsNetCore;
 		public static string EnvironmentName { get; set; }
@@ -81,16 +63,6 @@ namespace Clio.UserEnvironment
 			Settings = settingsRepository.GetEnvironment(options);
 		}
 
-		private static Version GetAppApiVersion() {
-			var apiVersion = new Version("0.0.0.0");
-			try {
-				string appVersionResponse = CreatioClient.ExecuteGetRequest(ApiVersionUrl).Trim('"');
-				apiVersion = new Version(appVersionResponse);
-			} catch (Exception) {
-			}
-			return apiVersion;
-		}
-
 		public string GetRegisteredPath() {
 			var environmentPath = Environment.GetEnvironmentVariable(PathVariableName);
 			string[] cliPath = (environmentPath?.Split(Path.PathSeparator));
@@ -115,25 +87,6 @@ namespace Clio.UserEnvironment
 
 		public string GetAssemblyFolderPath() {
 			return Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location);
-		}
-
-		public static void CheckApiVersion() {
-			var dir = AppDomain.CurrentDomain.BaseDirectory;
-			string versionFilePath = Path.Combine(dir, "cliogate", "version.txt");
-			var localApiVersion = new Version(File.ReadAllText(versionFilePath));
-			var appApiVersion = GetAppApiVersion();
-			if (appApiVersion == new Version("0.0.0.0")) {
-				MessageToConsole($"Your app does not contain clio API." +
-				 $"{Environment.NewLine}You should consider install it via the \'clio install-gate\' command.", ConsoleColor.DarkYellow);
-			} else if (localApiVersion > appApiVersion) {
-				MessageToConsole($"You are using clio api version {appApiVersion}, however version {localApiVersion} is available." +
-				 $"{Environment.NewLine}You should consider upgrading via the \'clio update-gate\' command.", ConsoleColor.DarkYellow);
-			}
-		}
-
-		public static void SetupAppConnection(EnvironmentOptions options) {
-			Configure(options);
-			CheckApiVersion();
 		}
 
 	}
